Fix several bugs in a4smedl grammar, parser, and codegen
Fixed several basic bugs to get mgen to the point where it can
successfully generate a full monitoring system from an architecture
file. The generated code most likely has bugs in it and may not even
compile, so there is more work to be done, but this is a major step
forward, at least.