Fix a bug with monitor state initialization.
If monitor state contained an opaque and there were any explicit
creation events for that monitor that set that opaque, the generated
code would not compile (due to using the string assignment function
instead of the opaque assignment function). This fixes that bug. There
is still a larger memory management issue here, though: see issue #55.