Remove "thread" type
The thread type was non-portable, forcing SMEDL monitors to depend on
pthreads which are not universally available. There was also no way to
safely implement threads as identities, since there is no comparison
operation, only equals. The thread type was rarely if ever used, and
there seemed to be no interest in keeping it, so this commit removes it.

Closes #41